Suggested correction are made in Syllabus API

+++ src/main/java/com/amitit/webapp/entity/Syllabus.java
@@ -21,7 +21,7 @@ public class Syllabus {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long sid;
+	private Integer sid;
 	
 	private String filePath;
 	private LocalDate uploadDate;
